requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
PyJWT[crypto]==2.9.0
httpx==0.28.1
google-genai==1.5.0
sqlalchemy==2.0.35
jinja2==3.1.4
flake8==7.1.1
pylint==3.3.1
python-dotenv==1.0.1
pytest==8.3.3
